Bug 1674902: Part 1 - s/uint32_t/long/ for return codes in mscom/Utils.h; r=Jamie
`HRESULT` is `typedef`'d as `long`, so this patch harmonizes the return codes.

Differential Revision: https://phabricator.services.mozilla.com/D95597

UltraBlame original commit: bf2fd0667eccdb609b3c2992cb47b765a656158c

+++ ipc/mscom/Utils.h
@@ -38,8 +38,8 @@ uintptr_t GetContainingModuleHandle();
 
 
 
-uint32_t CreateStream(const uint8_t* aBuf, const uint32_t aBufLen,
-                      IStream** aOutStream);
+long CreateStream(const uint8_t* aBuf, const uint32_t aBufLen,
+                  IStream** aOutStream);
 
 
 
@@ -48,7 +48,7 @@ uint32_t CreateStream(const uint8_t* aBuf, const uint32_t aBufLen,
 
 
 
-uint32_t CopySerializedProxy(IStream* aInStream, IStream** aOutStream);
+long CopySerializedProxy(IStream* aInStream, IStream** aOutStream);
 
 #if defined(MOZILLA_INTERNAL_API)
